requirements.txt
robotframework==6.0
robotframework-seleniumlibrary==6.1.0
selenium==4.10.0
robotframework-requests==0.9.4
requests==2.28.2
# สำหรับ Mobile
robotframework-appiumlibrary==1.5.0.6
Appium-Python-Client==2.11.0
